Anti-Bloat: Added yaml configuration for annotations mode per package
* Also split the yaml base class into a dedicated file, since it's not really for general use and makes it cleaner to
  see what is where.
