+++ nuitka/plugins/standard/ImplicitImports.py
@@ -15,7 +15,7 @@ import os
 from nuitka.__past__ import iter_modules, unicode
 from nuitka.importing.Importing import locateModule
 from nuitka.importing.Recursion import decideRecursion
-from nuitka.plugins.PluginBase import NuitkaYamlPluginBase
+from nuitka.plugins.YamlPluginBase import NuitkaYamlPluginBase
 from nuitka.utils.ModuleNames import ModuleName
 from nuitka.utils.Utils import isMacOS, isWin32Windows
 
@@ -641,7 +641,7 @@ According to Yaml 'overridden-environment-variables' configuration."""
         "wx._core",  # Too large generated code
         "pyVmomi.ServerObjects",  # Too large generated code
         "pyglet.gl",  # Too large generated code
-        "telethon.tl.types",  # Not performance relevant and slow C compile
+        # "telethon.tl.types",  # Not performance relevant and slow C compile
         "importlib_metadata",  # Not performance relevant and slow C compile
         "comtypes.gen",  # Not performance relevant and slow C compile
         "win32com.gen_py",  # Not performance relevant and slow C compile
